Move `_assocs` from `_module_to_kore` to `KDefinition`
